requirements/full.txt
beautifulsoup4==4.11.1
black==22.8.0
ephem==4.1.4
html2image==2.0.1
html2text==2020.1.16
Jinja2==3.1.2
Mastodon.py==1.8.0
mypy==1.0.0
nltk==3.7
numpy==1.23.3
pdoc==12.1.0
Pillow==9.2.0
playwright==1.45.1
pycmarkgfm==1.2.0
PyTumblr==0.1.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
selenium==4.4.3
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a27
tweepy==4.14.0
types-Pillow==9.2.1
types-PyYAML==6.0.11
types-requests==2.28.11.7
undetected-chromedriver==3.1.5.post4
Unidecode==1.3.6
unimport==0.11.3
